+++ docs/changelog.md
@@ -3,6 +3,8 @@
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format. 
 
 ## [Unreleased]
+### Fixed
+- Prevent duplication of CLI flags: [Issue 168](https://github.com/pepkit/looper/issues/168)
 
 ## [0.11.1] - 2019-04-17